StanPython/requirements.txt
fastapi
uvicorn
pystan
numpy<2.0.0
cython
matplotlib
scipy
pandas
pmdarima
